add 'load' command
The biggest change is probably the use of a new game_step() function,
that steps through 1 "iteration" of the game. We could just use
recursion and make game_loop() call itself to progress into the game,
but the iterative way is a different flavor worth trying out (as opposed
to NW, which uses recursion).